+++ docs/HISTORY.rst
@@ -4,20 +4,26 @@ Changelog
 1.0.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Move controlpanel to z3c.form. Removes dependency on CMFDefault and formlib.
+  [pbauer]
 
 
 1.0.2 (2015-06-04)
 ------------------
 
 - use the recommended way of including Products.CMFCore permissions
-  in configure.zcml so it works for all Plone versions [jcerjak]
+  in configure.zcml so it works for all Plone versions
+  [jcerjak]
+
 - Remove dependency on portal_factory (Plone 5 compatibility)
   [mattss]
 
+
 1.0.1 - 2012-02-10
 ------------------
+
 - preloading Products.CMFCore.permissions.zcml
+
 
 1.0.0 - 2010-12-03
 ------------------